refactor: Github Copilot Rewied applied
